fix: 날짜를 ISO8601 형식으로 변환하지 않는 문제 수정
